Working first version of loop exit pass
Added code to AnnotateLoopEnds which inserts
a boogie 'assert false' at any exit of a loop.

The assertions are inserted at the beginning
of the "loop exit" basic block, that is the
immediate node outside of the loop.
In order to prevent a random failure, the pass
now requires the LoopSimplify pass to run first,
which guarentees that each loop exit block is
dominated only by the loop.